횡단관심사 관련 파일 - common 폴더로 이동, swagger 설정
