+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
       dockerfile: Dockerfile.local
     restart: always
     env_file:
-      - ./src/config/env/.local.env
+      - ./src/common/config/env/.local.env
     volumes:
       - ./src:/app/src
       - ./test:/app/test
