+++ nest-cli.json
@@ -6,7 +6,7 @@
     "deleteOutDir": true,
     "assets": [
       {
-        "include": "./config/env/*.env",
+        "include": "./common/config/env/*.env",
         "outDir": "./dist"
       }
     ]
