+++ package.json
@@ -26,6 +26,7 @@
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/passport": "^10.0.3",
     "@nestjs/platform-express": "^10.3.1",
+    "@nestjs/swagger": "^7.3.0",
     "@prisma/client": "^5.8.1",
     "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
