+++ src/project/dto/get-projects.response.dto.ts
@@ -1,4 +1,4 @@
-import { PageResponseDto } from 'src/pagination/dto'
+import { PageResponseDto } from 'src/common/pagination'
 import { ProjectSummaryDto } from './project-summary.dto'
 
 export class GetProjectsResponseDto extends PageResponseDto<ProjectSummaryDto> {}
